+++ src/xrGame/game_sv_base.cpp
@@ -280,7 +280,7 @@ struct player_exporter
 
 		game_PlayerState*	curr_ps = tmp_client->ps;
 		
-		u16 tmp_flags = curr_ps->flags__;
+		u32 tmp_flags = curr_ps->flags__;
 		if (to_ps == curr_ps)	
 			curr_ps->setFlag(GAME_PLAYER_FLAG_LOCAL);
 
@@ -327,7 +327,7 @@ void game_sv_GameState::net_Export_Update(NET_Packet& P, ClientID id_to, ClientI
 	game_PlayerState* A			= get_id(id);
 	if (A)
 	{
-		u16 bk_flags			= A->flags__;
+		u32 bk_flags			= A->flags__;
 		if (id==id_to)	
 		{
 			A->setFlag(GAME_PLAYER_FLAG_LOCAL);
